v1.2.9: Fix TA-Lib errors, add CoinGecko mappings, and add Logs tab
**Critical Fixes:**
1. **TA-Lib Data Type Error** - Fixed "input array type is not double" error
   - Converted all numpy arrays to float64 before passing to TA-Lib functions
   - Affects RSI, MACD, Bollinger Bands, Moving Averages, and Volume SMA
   - This should restore screener functionality and technical indicator calculations

2. **CoinGecko Coin Mappings** - Added missing coin ID mappings
   - Added 20 missing coin mappings (ATOM, NEAR, APT, SUI, UNI, AAVE, POL, ARB, OP, RENDER, FET, GRT, PEPE, DOGE, LTC, BCH, ETC, TIA, INJ)
   - Reduces "No CoinGecko mapping" warnings from 19 to 0

3. **CoinGecko Rate Limiting** - Reduced API call frequency
   - Decreased rate limiter from 10 to 5 calls/minute
   - Increased cache time from 60 to 120 minutes
   - Should significantly reduce 429 "Too Many Requests" errors

**New Feature:**
4. **Dedicated Logs Tab** - Professional log viewer with features:
   - Shows last 2 days of logs (newest first)
   - Color-coded by severity:
     * Red background - Errors
     * Orange text - Warnings
     * Orange background - Bot starts (easy identification)
     * Blue text - Info messages
   - Filters: "Show Errors Only" and "Show Warnings" checkboxes
   - Auto-refresh every 5 seconds (can pause/resume)
   - Dark theme terminal-style display
   - Proper HTML escaping for security

**Impact:**
- Screener should now find trading opportunities (was finding 0)
- Technical indicators will calculate properly
- Claude AI will receive complete market data
- Significantly fewer CoinGecko API errors
- Easy monitoring of bot activity and errors

+++ src/signals.py
@@ -36,7 +36,7 @@ class SignalGenerator:
             RSI series
         """
         period = period or self.config.get("rsi_period", 14)
-        return talib.RSI(df['close'].values, timeperiod=period)
+        return talib.RSI(df['close'].values.astype(np.float64), timeperiod=period)
 
     def calculate_macd(self, df: pd.DataFrame) -> tuple:
         """
@@ -53,7 +53,7 @@ class SignalGenerator:
         signal = self.config.get("macd_signal", 9)
 
         macd, macd_signal, macd_hist = talib.MACD(
-            df['close'].values,
+            df['close'].values.astype(np.float64),
             fastperiod=fast,
             slowperiod=slow,
             signalperiod=signal
@@ -75,7 +75,7 @@ class SignalGenerator:
         std = self.config.get("bb_std", 2.0)
 
         upper, middle, lower = talib.BBANDS(
-            df['close'].values,
+            df['close'].values.astype(np.float64),
             timeperiod=period,
             nbdevup=std,
             nbdevdn=std
@@ -96,8 +96,8 @@ class SignalGenerator:
         short_period = self.config.get("ma_short", 50)
         long_period = self.config.get("ma_long", 200)
 
-        ma_short = talib.SMA(df['close'].values, timeperiod=short_period)
-        ma_long = talib.SMA(df['close'].values, timeperiod=long_period)
+        ma_short = talib.SMA(df['close'].values.astype(np.float64), timeperiod=short_period)
+        ma_long = talib.SMA(df['close'].values.astype(np.float64), timeperiod=long_period)
 
         return ma_short, ma_long
 
@@ -152,7 +152,7 @@ class SignalGenerator:
             df['ma_short'], df['ma_long'] = self.calculate_moving_averages(df)
 
             # Volume SMA
-            df['volume_sma'] = talib.SMA(df['volume'].values, timeperiod=20)
+            df['volume_sma'] = talib.SMA(df['volume'].values.astype(np.float64), timeperiod=20)
 
             return df
